function_source/solver/requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
pulp